Dedup tests.
